requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.5
python-socketio==5.10.0
requests==2.31.0
APScheduler==3.10.4
PyYAML==6.0.1
watchdog==3.0.0
gunicorn==21.2.0